Close #24 : Test général - Fixation de divers bugs - Divers ajouts/modifications - Adaptation mobile
