+++ .gitignore
@@ -45,3 +45,4 @@
 
 .idea
 /composer.phar
+/projet/
