+++ app/Resources/views/figure/form_figure.html.twig
@@ -112,6 +112,11 @@
 
             </div>
 
+            <div id="line_break_xs">
+                <br>
+                <br>
+            </div>
+
             <div class="col-sm-6">
                 <div class="form-group">
