+++ app/Resources/views/login/account.html.twig
@@ -18,7 +18,7 @@
         </legend>
 
         <div class="row">
-            <div class="col-sm-6">
+            <div class="col-lg-6">
                 <div class="form-group group_login">
 
                     {{ form_row(formUser.username) }}
@@ -44,10 +44,10 @@
                 </div>
             </div>
 
-            <div class="col-sm-6">
+            <div class="col-lg-6">
 
                 <div class="user_avatar">
-                    <img src="{{ asset(user.avatar.webPath)}}" alt="{{ user.avatar.alt }}" width="250" height="180">
+                    <img src="{{ asset(user.avatar.webPath)}}" alt="{{ user.avatar.alt }}" width="210" height="150">
                 </div>
 
                 <div class="form-group group_login">
