+++ app/Resources/views/login/signup.html.twig
@@ -18,7 +18,7 @@
         </legend>
 
         <div class="row">
-            <div class="col-sm-6">
+            <div class="col-lg-6">
                 <div class="form-group group_login">
 
                     {{ form_row(formUser.username) }}
@@ -44,7 +44,7 @@
                 </div>
             </div>
 
-            <div class="col-sm-6">
+            <div class="col-lg-6">
                 <div class="form-group group_login">
 
                     {{ form_label(formUser.avatar, 'Votre avatar :', {'label_attr' : {'class' : 'control-label'}}) }}
