+++ app/Resources/views/menu.html.twig
@@ -1,6 +1,25 @@
 <nav class="navbar navbar-default">
 
+    <div class="navbar-header">
+
+
+            <h1 id="title_xs" class="col-xs-10">SnowTricks</h1>
+
+            <button type="button" class="navbar-toggle col-sm-2" data-toggle="collapse" data-target=".navbar-collapse">
+
+                <span class="icon-bar"></span>
+
+                <span class="icon-bar"></span>
+
+                <span class="icon-bar"></span>
+
+            </button>
+
+        </div>
+
+
     <div class="collapse navbar-collapse">
+
 
         <ul class="nav navbar-nav">
 
@@ -29,7 +48,7 @@
 
             <li><a href="{{ path('lexique') }}">Lexique</a></li>
             <li><a href="{{ path("contact") }}">Contact</a></li>
-            <li><a href="{{ path("mentions") }}">Mention légales</a></li>
+            <li><a href="{{ path("mentions") }}">Mentions légales</a></li>
 
         </ul>
