+++ app/config/config.yml
@@ -66,7 +66,6 @@ swiftmailer:
     transport: "%mailer_transport%"
     auth_mode: "%mailer_auth_mode%"
     host:      "%mailer_host%"
-    port:      "%mailer_port%"
     username:  "%mailer_user%"
     password:  "%mailer_password%"
     spool:     { type: memory }
