+++ app/config/routing_dev.yml
@@ -12,3 +12,7 @@ _errors:
 
 _main:
     resource: routing.yml
+
+coresphere_console:
+    resource: .
+    type: extra
