+++ composer.json
@@ -32,7 +32,8 @@
     },
     "require-dev": {
         "sensio/generator-bundle": "^3.0",
-        "symfony/phpunit-bridge": "^3.0"
+        "symfony/phpunit-bridge": "^3.0",
+        "coresphere/console-bundle": "0.5.0"
     },
     "scripts": {
         "symfony-scripts": [
