+++ src/AppBundle/Entity/Image.php
@@ -39,7 +39,7 @@ class Image
     /**
      * @var \DateTime
      *
-     * @ORM\Column(name="dateCreate", type="datetime")
+     * @ORM\Column(name="date_create", type="datetime")
      */
     private $dateCreate;
