+++ src/AppBundle/Entity/Video.php
@@ -53,7 +53,7 @@ class Video
     /**
      * @var \DateTime
      *
-     * @ORM\Column(name="dateCreate", type="datetime")
+     * @ORM\Column(name="date_create", type="datetime")
      */
     private $dateCreate;
