+++ web/css/error.css
@@ -1,0 +1,4 @@
+.error{
+    text-align: center;
+    font-weight: bolder;
+}
