+++ web/css/figure.css
@@ -14,16 +14,16 @@
     text-align: left;
 }
 
-#video_view{
+#video_view {
     float: left;
 }
 
-#carousel{
+#carousel {
     margin: auto;
     width: 80%;
 }
 
-.carousel-inner img{
+.carousel-inner img {
     margin: auto;
 }
 
@@ -46,44 +46,44 @@ legend {
     margin-bottom: 1%;
 }
 
-#comment_form_figure{
+#comment_form_figure {
     margin-bottom: 5%;
 }
 
-#appbundle_comment_content{
+#appbundle_comment_content {
     resize: vertical;
 }
 
-#comments_post_figure{
+#comments_post_figure {
     text-align: left;
     border: 0.2em solid black;
     border-radius: 0.3em;
 }
 
-.comment_figure{
+.comment_figure {
     border-top: 0.05em solid black;
     margin: 0 3% 4% 3%;
 
 }
 
-.comment_auteur_figure{
+.comment_auteur_figure {
     color: #0088CC;
     font-weight: bold;
     margin-bottom: 0;
 }
 
-.comment_auteur_figure > em{
+.comment_auteur_figure > em {
     color: black;
     font-style: italic;
     font-weight: normal;
     font-size: smaller;
 }
 
-.comment_content_figure{
+.comment_content_figure {
     margin-bottom: 1%;
 }
 
-#btn_valid_form{
+#btn_valid_form {
     padding: 1%;
     margin-bottom: 1%;
     margin-right: 1%;
@@ -91,7 +91,7 @@ legend {
     font-size: larger;
 }
 
-#comments_post_figure > span{
+#comments_post_figure > span {
     padding: 0 1% 0 1%;
     background-color: white;
     color: #0088CC;
@@ -102,14 +102,14 @@ legend {
     left: 0.5em;
 }
 
-#comments_empty_figure{
+#comments_empty_figure {
     text-align: center;
 }
 
-#comment_pagination{
+#comment_pagination {
     text-align: center;
 }
 
-.comment_avatar{
-    text-align: center;
+.comment_avatar {
+    margin-bottom: 0.5em;
 }
