+++ web/css/layout.css
@@ -2,7 +2,7 @@ body {
     padding: 1% 3% 5% 3%;
 }
 
-#title {
+#title_lg {
     font-weight: bold;
     font-size: 48px;
     border-top: 2px solid black;
@@ -14,11 +14,48 @@ body {
     margin-bottom: 1%;
 }
 
+@media (max-width: 640px) {
+    #title_lg {
+        display: none;
+    }
+}
+
+#title_xs {
+    font-weight: bold;
+    font-size: 39px;
+    color: white;
+    background-color: #0088CC;
+    text-align: center;
+    margin: auto;
+    padding: 1% 1%;
+
+}
+
+@media (min-width: 640px) {
+    .navbar-header {
+        display: none;
+    }
+}
+
+.navbar-header {
+    background-color: #0088CC;
+    border-top: 2px solid black;
+}
+
+.navbar-toggle {
+    margin-right: 0.4em;
+    background-color: white;
+    padding: 0.5em;
+}
+
 .navbar-nav {
     text-align: center;
+    float: none;
 }
 
 .navbar-nav > li {
+    display: inline-block;
+    float: none;
 }
 
 .navbar-default .navbar-nav > li > a, .dropdown-menu > li > a {
@@ -43,6 +80,7 @@ body {
 
 .panel-body {
     padding: 0;
+    margin-bottom: 1.2em;
 }
 
 .panel-title {
@@ -50,17 +88,18 @@ body {
     font-size: 1.2em;
 }
 
+.news_info {
+    padding: 1%;
+    margin: 0.5em;
+}
+
 .news_comment {
     padding-right: 0;
 }
 
-.news_info, .news_media {
-    padding: 1%;
-}
-
 .news_name {
     font-size: 1.2em;
-    margin: 0;
+    margin-bottom: 0;
 }
 
 .news_date {
@@ -73,6 +112,23 @@ body {
     padding: 2% 0 0 0;
 }
 
+@media (max-width: 640px) {
+    #login_user_sm {
+        display: none;
+    }
+
+    #login_user_xs{
+        margin-top: -0.5em;
+        margin-bottom: 0.6em;
+    }
+}
+
+@media (min-width: 641px) {
+    #login_user_xs {
+        display: none;
+    }
+}
+
 #btn_user {
     text-align: center;
 }
@@ -82,11 +138,6 @@ body {
     color: #0088CC;
     background-color: white;
     border-color: black;
-    padding: 2.5% 2.5% 2.5% 6%;
-    border-bottom-left-radius: 22%;
-    border-bottom-right-radius: 0;
-    border-top-left-radius: 22%;
-    border-top-right-radius: 0;
     font-weight: bolder;
 }
 
@@ -95,12 +146,35 @@ body {
     color: #0088CC;
     background-color: white;
     border-color: black;
-    padding: 2.5% 6% 2.5% 2.5%;
-    border-bottom-left-radius: 0;
-    border-bottom-right-radius: 22%;
-    border-top-left-radius: 0;
-    border-top-right-radius: 22%;
     font-weight: bolder;
+}
+
+@media (max-width: 1025px) {
+    .btn_login_left {
+        border-radius: 22%;
+    }
+
+    .btn_login_right {
+        border-radius: 22%;
+    }
+}
+
+@media (min-width: 1025px) {
+    .btn_login_left {
+        padding: 2.5% 2.5% 2.5% 6%;
+        border-bottom-left-radius: 22%;
+        border-bottom-right-radius: 0;
+        border-top-left-radius: 22%;
+        border-top-right-radius: 0;
+    }
+
+    .btn_login_right {
+        padding: 2.5% 6% 2.5% 2.5%;
+        border-bottom-left-radius: 0;
+        border-bottom-right-radius: 22%;
+        border-top-left-radius: 0;
+        border-top-right-radius: 22%;
+    }
 }
 
 #btn_user a:hover {
@@ -118,4 +192,27 @@ body {
 .lexique h3 {
     font-size: x-large;
     font-weight: bold;
+}
+
+#news{
+    margin-bottom: 0.6em;
+}
+
+@media (max-width: 640px){
+    #home_list_figures{
+        padding: 0;
+    }
+
+    #home_list_figures > .col-lg-6{
+        padding: 0;
+    }
+
+    .home_figure{
+        margin-bottom: 0.6em;
+    }
+
+    .home_figure > div {
+        padding: 0;
+    }
+
 }
